main.c en cours

+++ libasm/libasm.h
@@ -6,7 +6,7 @@
 /*   By: lduhamel <lduhamel@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/13 11:44:24 by lduhamel          #+#    #+#             */
-/*   Updated: 2020/10/18 17:25:16 by lduhamel         ###   ########.fr       */
+/*   Updated: 2020/10/21 14:41:21 by lduhamel         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@
 # include <stdio.h>
 # include <stdlib.h>
 
-int     ft_strlen(char *rdi);
+ssize_t     ft_strlen(char *rdi);
 char	*ft_strcpy(char *rdi, const char *rsi);
 int		ft_strcmp(const char *rdi, const char *rsi);
 ssize_t	ft_write(int rdi, const void *rsi, size_t rdx);
